+++ EssentialFeedJorgeiOSTests/Controllers/FeedViewControllerTests.swift
@@ -361,37 +361,3 @@ final class FeedViewControllerTests: XCTestCase {
         }
     }
 }
-
-
-// MARK: - Extension. FeedImageTableViewCell
-private extension FeedImageTableViewCell {
-    var isShowingLocation: Bool {
-        return !locationContainer.isHidden
-    }
-    
-    var isShowingImageLoadingIndicator: Bool {
-        return feedImageContainer.isShimmering
-    }
-    
-    var locationText: String? {
-        return locationLabel.text
-    }
-    
-    var descriptionText: String? {
-        return descriptionLabel.text
-    }
-    
-    var renderedImage: Data? {
-        return feedImageView.image?.pngData()
-    }
-    
-    var isShowingRetryAction: Bool {
-        return !feedImageRetryButton.isHidden
-    }
-    
-    
-    func simulateRetryAction() {
-        feedImageRetryButton.simulateTap()
-    }
-}
-
