+++ EssentialFeedJorgeiOSTests/Helpers/FeedImageCell+TestHelpers.swift
@@ -1,0 +1,43 @@
+//
+//  FeedImageCell+TestHelpers.swift
+//  EssentialFeedJorgeiOSTests
+//
+//  Created by jfuentescasillas on 15/09/23.
+//
+
+
+import UIKit
+import EssentialFeedJorgeiOS
+
+
+// MARK: - Extension. FeedImageTableViewCell
+extension FeedImageTableViewCell {
+    var isShowingLocation: Bool {
+        return !locationContainer.isHidden
+    }
+    
+    var isShowingImageLoadingIndicator: Bool {
+        return feedImageContainer.isShimmering
+    }
+    
+    var locationText: String? {
+        return locationLabel.text
+    }
+    
+    var descriptionText: String? {
+        return descriptionLabel.text
+    }
+    
+    var renderedImage: Data? {
+        return feedImageView.image?.pngData()
+    }
+    
+    var isShowingRetryAction: Bool {
+        return !feedImageRetryButton.isHidden
+    }
+    
+    
+    func simulateRetryAction() {
+        feedImageRetryButton.simulateTap()
+    }
+}
